ShelfCam-Backend-main/requirements_clean.txt
fastapi==0.115.13
uvicorn==0.34.3
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
bcrypt==4.3.0
ultralytics
opencv-python
torch
torchvision